fix warns

+++ src/managers/LayoutManager.cpp
@@ -12,7 +12,7 @@ IHyprLayout* CLayoutManager::getCurrentLayout() {
 void CLayoutManager::switchToLayout(std::string layout) {
     for (size_t i = 0; i < m_vLayouts.size(); ++i) {
         if (m_vLayouts[i].first == layout) {
-            if (i == m_iCurrentLayoutID)
+            if (i == (size_t)m_iCurrentLayoutID)
                 return;
 
             getCurrentLayout()->onDisable();
